Implement basic styling for multi-symbol buttons
Summary:
This diff adds some of the basic styling elements and component model for the multi-symbol buttons.
I choose to try and encapsulate all of the multi-symbol buttons in a single
`MultiSymbolKeypadButton` (need to revisit all of these names), which seems doable.

{F483302}

All of the long-press handling remains a TODO.

Getting my hands dirty with Flexbox here, so would love feedback on how I made the layouts happen!

On Asana @ https://app.asana.com/0/114493820703878/115226565502040.

Test Plan:
- Open `index.html`.
- Enable the `MULTI_BUTTON` keypad.
    - Verify that the leftmost button displays just the primary symbol, with the corner decal.
    - Verify that the middle button displays the primary symbol and two secondary symbols, with the corner decal.
    - Verify that the rightmost button displays the primary symbol and one secondary symbols, with the corner decal.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D26740