Add basic popover
Summary:
This diff adds the basic popover that will display a number of buttons. Still no gestural support
though--just some of the layout / information architecture that we'll need.

On Asana @ https://app.asana.com/0/114493820703878/115507156056144.

Test Plan:
Add a `MultiSymbolPopover` to one of the keypads. Verify that it appears as vertical columnar modal above the key, with some shadow and the correct keys.

{F483434}

(Need to figure out why the top border is twice as wide--will be addressed in a polish pass later on.)

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D26753