Allow keypad to be configured with arbitrary extra symbols
Summary:
This diff adds support for the 'extra symbols' button that will encapsulate any
additional variables and symbols (like Pi) that might only be necessary on a
per-input basis.

As Perseus will configure the keypad with both its type and any extra symbols
in one call, these parameters are packaged together as a `configuration` object
that is passed to the parent keypad, with some of those objects then being
passed selectively down to the children.

On Asana @ https://app.asana.com/0/114493820703878/116028938008310.

Test Plan:
- Open `index.html`.
- Choose the `TEST_MULTI_SYMBOL` keypad.
    - Verify that the rightmost symbol displays four symbols, including Pi.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D26808