Create "Advanced" Expression keypad
Summary:
This diff ties in all the functionality from the last few diffs to build out the entire "Advanced"
Expression keypad. Some of the buttons in the second page are non-functional, and the gestures
haven't been tied in yet (since those diffs are still in their early stages), but the presentation
is there.

On Asana @ https://app.asana.com/0/114493820703878/114997243483975.

Test Plan:
- Open `index.html`.
- Click the `ADVANCED_EXPRESSION` keypad.
    - Verify that the buttons match the mocks, in both panes.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D26812