Layout Expression keypads by column rather than row
Summary:
The Expression keypads are designed such that the buttons are semantically grouped by column, rather
than by row. As such, it's easier to lay them out as grouped by column. This also makes it simple
to vary the sizes of the columns, which is necessary for the second page of both variants of the
Expression keypad.

We may want to re-organize all keypads to group their keys by row, though it doesn't seem totally
necessary right now for the non-Expressions.

On Asana @ https://app.asana.com/0/114493820703878/116270775238797.

Test Plan:
- Open `index.html`.
- Click through each keypad.
    - Verify that it's layed out correctly, even on resizing.
- Open the second page of the `BASIC_EXPRESSION` and `ADVANCED_EXPRESSION` keypads.
    - Verify that in both cases, the rightmost column is the same width as the columns on the first page, while the remaining columns are equal-width, taking up the remaining space.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D26816