Polish and simplify styles and layout
Summary:
This is a polish pass over some of our styles and layouts. It pulls in some of the pixel-perfect
designs from Zeplin, and also refactors the layout logic in our components, which was getting
messy and can now be simplified given our use of SVG.

In addition, I added proper layout for the 'many symbol' button using text rather than SVG. It's
unclear if this will be the appropriate long-term solution, so I didn't re-add all of the symbol
plumbing in this diff.

Test Plan:
- Open the advanced Expression keypad.
- Marvel at the slick designs.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D26907