Better algorithm for determining where to place the cursor when dragging the handle
Summary:
Think of the handle as a gun that shooting tiny little bullets vertically
up the page.  pew! pew!  As the bullets travel if they hit anything of
importance like a leaf node or a numerator, denominator, fraction, etc. it
ask MathQuill to position the cursor beside that node and then we will update
the position of the handle to be below the cursor.

If the tiny bullets reach the top of the math-input container before hitting
anything, they'll bounce off the top and head back down the page.  pew! pew!

If they reach the bottom and they haven't hit anything then we finally check
if touch events we're getting are withing the left border of the container
at which point we'll ask MathQuill to position the cursor at the end at which
point we'll up date the position of the handle.

Test Plan:
- type some math with exponents, fractions, parens, and roots
- click to make the handle appear
- drag the handle around to position it anywhere within the expression

Reviewers: charlie

Reviewed By: charlie

Differential Revision: https://phabricator.khanacademy.org/D26956