Handle all key presses in reducers
Summary:
In our previous design, each key had an `onClick` that dispatched some action on the store. Now,
we move the logic of handling the triggering of a key into the store. This will simplify the
process of triggering an action later on, when our gesture management system needs to be able to
trigger actions knowing only the identifier of the node that was touched. It also consolidates
this logic into one place (our reducers).

Depends On: D26957

Test Plan:
- Open the keypads.
- Verify that clicking on a numeral or math command triggers the appropriate action in the math input.
- Verify that clicking the ellipses in the Expression keypads takes you to the second page.
- Verify that clicking on a math command in the second page takes you to the first page.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D26958