Add echo state animation on press
Summary:
This diff adds in the echo animation on press using `ReactCSSTransitionGroup`.
The echoes are managed globally, rather than local to the component, as we may
need to display them even after the DOM nodes of the buttons that were pressed
have been removed from the DOM (e.g., in the event of a popover button being
clicked).

On Asana @ https://app.asana.com/0/114493820703878/116075542842165.

Test Plan:
- Open up a keypad.
- Click a button.
    - Verify that the echo animation appears, moves up, and fades out.
- Click a button, and click it again before the echo state is done animating.
    - Verify that there are two echoes on screen at the same time, for the same key.
- Click on a multi-touch button.
    - Verify that the echo state reflects the default action.
- Open a multi-page keypad; go to the second page and click on a button.
    - Verify that the keypad swipes back to the first page, but the echo remains fixed horizontally.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D26990