Substitute placeholder symbols for second Expression pane
Summary:
This diff adds in some placeholder iconography for the second page of icons, so they at least
render something reasonable. We will likely replace these in the near future.

Test Plan:
Take a gander at the second page of an Expression keypad, and verify that every button has an icon.

{F517494}

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27023