Render shadow over keypad when popover is focused
Summary:
This diff renders a shadow behind the popover, but above the keypad, whenever the popover is
focused. This behavior should've been trivial to implement; however, due to some issues with the
simultaneous usage of both `transform` and `z-index`, it became quite complicated to stack the
layers correctly. As such, the shadow is rendered in a non-ideal way.

@kevinb and I discussed this offline, and an alternative would be to render the popovers in the same
way that we render the echoes (absolute, relative to the keypad, rather than relative to the
buttons over which they appear). I whipped up a diff to that effect, but there was so much
additional complexity that it didn't feel worth it. If some additional behavior makes this current
approach infeasible, we can revisit; but this seems simple and expedient.

On Asana @ https://app.asana.com/0/114493820703878/117544137365087.

Test Plan:
- Open up an Expression keypad.
- Hold on a popover node.
    - Verify that all the keys behind the popover darkens when the popover appears.
    - Verify that the popover is rendered above the shadow.
    - Verify that the sidebar receives the darkening effecft as well.

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27081