Separate pagination data from keypad configuration
Summary:
In preparation for the introduction of a bunch of new pagination-related data (to implement
swiping behavior), this diff separates out the base configuration of a keypad from its current
page.

This split will become more useful when the `pagerReducer` also includes the x-offset and houses
the logic for deciding on which page the keypad should land after a swipe gesture has ended.

Test Plan: NFC. Open up the keypads, swipe around. Change keypads, change pages. Everything should be as it was.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27082