Implement basic dismissal behavior
Summary:
This diff implements basic 'dismiss' functionality. Right now, the keypad will slide in on load,
slide out on clicking 'dismiss', and slide in again whenever you click on a keypad with the
selector (but not if you switch keypad types while a keypad is active). This avoids doing any
additional work to manage dismissal, as the plan is to enable keypads by calling `ConfigureKeypad`
from the Perseus bridge anyway, so it mirrors the planned data flow.

Unfortunately, this don't manage the case in which you re-active the keypad while its in the middle
of dismissing itself (in which case the keypad should stop dismissing and start sliding back up
into place). That kind of effect isn't possible with CSS animations. Instead, we'll need to use a
Spring or perhaps `react-motion`.

On Asana @ https://app.asana.com/0/114493820703878/118096062690058.

Test Plan:
(This test plan relies on the "keypad type selector" being re-enabled.)

- Open up the playground.
    - Verify that the keypad slides into place.
- Click 'dismiss'.
    - Verify that the keypad slides out of view.
- Click on the selection button for a different keypad.
    - Verify that the new keypad slides into place.
- Click on the selection button for a different keypad without dismissing
    - Verify that the new keypad appears, but that there is no slide animation.
- Dismiss and click on the selection button for the dismissed keypad.
    - Verify that it slides back up into place.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27178