Make view pagination UI a configurable setting
Summary:
This diff makes the view pagination UI a configurable setting. Users can choose to use any of:

- The view page indicators at the bottom.
- The tab bar indicator at the top.
- The toggle button in the right pane.

In the event that the view page or tab bar indicators are selected, the toggle button becomes a
'Move Left' key, as per the mocks.

Test Plan: Configure a keypad (at `/custom.html`) with each of the three settings. Verify that they match the mocks.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27143