Implement naive `TOGGLE_SIGN` operator
Summary:
This diff implements the `TOGGLE_SIGN` operator. The implementation is incredibly naive in that it
won't work properly for keypads that have selection enabled, and it won't work properly for keypads
that might yield nested expressions, since it just inserts the minus sign at the front of the input
or removes it if it's at the front of the input. But it should serve our purposes for usability
testing.

On Asana @ https://app.asana.com/0/114493820703878/120941711043335.

Test Plan:
- Open up the Fraction keypad.
- Type some math.
- Hit +- button.
    - Verify that a - appears at the front, but the cursor remains in place.
- Hit +- button.
    - Verify that the - disappears from the front, and the cursor remains in place.
- Test the following cases:
    - Cursor at the front, no - present.
    - Cursor at the front, - present.
    - Cursor at the -.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27154