Tweak echo animation timing function and make animations configurable
Summary:
For user testing, we want to try out two variants of the echo animation. In the first, the echo
state both slides up and fades out, using a Bezier curve of (0, 1, 0.25, 1) over 800ms for the
slide and (1, 0, 0.9, 1) over 400ms for the fade. In the second, the echo only fades, with a Bezier
curve of (0.25, 0, 0.75, 1) over 300ms.

This diff implements those two animations and makes them configurable via the `custom.html` UI.

On Asana @ https://app.asana.com/0/114493820703878/130808983737528 and https://app.asana.com/0/114493820703878/130808983737530.

Test Plan:
- Open up `custom.html`.
- Choose `SLIDE_AND_FADE`.
    - Verify that the echo animation slides up and fades out, much quicker than before, and with some non-linear effects.
- Open up `custom.html`.
- Choose `FADE_ONLY`.
    - Verify that the echo animation only fades out (no slide effect), and even quicker than in the `SLIDE_AND_FADE` case.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27177