Trigger two deletions when deleting a >= or <= symbol
Summary:
MathQuill represents greater-than-or-equal-to and less-than-or-equal-to as ligatures that combine
the greater-than or less-than and equal-to symbols, respectively. Thus, when you hit backspace
after one of those symbols, it morphs them into greater-than or less-than, whereas we want to delete
the entire symbol. This diff modifies our deletion behavior to trigger two backspaces when deleting
those symbols.

Note that 'not-equal-to' is _not_ treated in the same way--it is deleted fully with a single backspace.

On Asana @ https://app.asana.com/0/114493820703878/130801370525426.

Test Plan:
- Type some math, followed by a >=.
- Hit delete.
    - Verify that the >= is fully deleted, and the rest of the math is intact.
- Type some math, followed by a <=.
- Hit delete.
    - Verify that the <= is fully deleted, and the rest of the math is intact.
- Type some math, followed by a =/=.
- Hit delete.
    - Verify that the =/= is fully deleted, and the rest of the math is intact.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27176