Guard against deleting multiple sets of parens
Summary:
In our parens deletion algorithm, when hitting backspace with our cursor on the left edge of a set
of parens, we were checking if the node to the left of the parens was a command (like \log or \sin)
so that we could also include that command in our selection or deletion. However, if we had adjacent
sets of parens, then the node to the left would _appear_ to be a valid command, as it took the form
\left(, even though we didn't actually want to delete it.

This diff adds a new test to account for that case, and fixes it by guarding against deleting parens
that are to the left of a set of empty parens.

@kevinb's code was too robust!

On Asana @ https://app.asana.com/0/114493820703878/130801370525424.

Test Plan:
- Type in two sets of parens.
- Place your cursor in the second empty set.
- Hit 'Delete'.
    - Verify that only the first set of parens is deleted.
- Type in `cos()` and try to delete with your cursor in the empty parens.
    - Verify that the entire expression is deleted.
- Run all the unit tests, which give us very good coverage here.

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27175