Allow `MathInput` to take an initial value and callback when it changes
Summary:
As with the other Perseus inputs, `MathInput` needs to provide an `onChange` callback to notify its
parent when its value changes (so as to allow it to be graded properly). The design I opted for
here mirrors that of the other Perseus controlled input components in that `value` is stored as a
prop rather than as state, with the expectation being that it is passed back down to the component
by the parent. See Perseus's own `MathInput` for comparison.

Alternatively, we could store the input value as state and only accept an `initialValue` as props.

Test Plan: Coming in the integration diff. For now: set an initial value in `app.js` and verify that the input starts with that value.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27282