Add class-based namespacing to `overrides.css`
Summary:
This ensures that the rules are more specific than `mathquill.css`, but won't be applied to non-keypad
math inputs in Perseus.

On Asana @ https://app.asana.com/0/114493820703878/133129670297602.

Test Plan: Coming in the integration diff. For now, verify that the cursor and selection rect are styled properly in the playground.

Reviewers: kevinb

Reviewed By: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27324