Render selected text with white border
Summary:
MathQuill renders some of its math using borders, rather than true text. For example, the top bar
in the radical symbol is rendered as a top border. As such, we need to set the border color to
white when rendering the selected text; otherwise, that part of the math will be improperly colored
(before, it was invisible, since we were setting a transparent border).

On Asana @ https://app.asana.com/0/114493820703878/129252107834882.

Test Plan:
- Open up the keypad.
- Type in a radical; click past it and hit 'Delete'.
    - Verify that the top of the radical is white, matching the rest of the selected text.
- Make sure any other selections render as before (parens, fractions, etc.)

Auditors: kevinb