Don't blur input on scroll
Summary:
Previously, the input was being blurred whenever a touchstart event occurred with a target outside
of the input itself. Thus, even if you scrolled (e.g., to get the input into view or see part of
the problem), the input would blur and the keypad would disappear. This made for an exceptionally
frustrating user experience, since you couldn't scroll around to read the problem that you were
trying to answer without re-focusing the input after every scroll.

Instead, we now track the page scroll and see if it changed over the course of the touch event.
This isn't super robust (since we're just looking at the start and end `scrollTop`) but should make
for a better user experience.

On Asana @ https://app.asana.com/0/114493820703878/133272025997757.

Test Plan:
- Open up an exercise with multiple inputs in webapp ([link](http://localhost:8080/math/cc-fourth-grade-math/cc-4th-mult-div-topic/cc-4th-area-models/e/multiplying-2-digit-numbers-with-area-models)).
- Tap outside and scroll around.
    - Verify that the focused input does not blur.
- Tap outside without scrolling.
    - Verify that the focused input blurs.
- Tap into an input while an input is focused.
    - Verify that it immediatley focuses the new input.

Reviewers: benkomalo

Reviewed By: benkomalo

Subscribers: kevinb

Differential Revision: https://phabricator.khanacademy.org/D27359